[1415] Spinta sync improvements: Refactor current implementation
- Checkpoint 3: Move logic to controllers
