+++ spinta/cli/helpers/sync/__init__.py
@@ -1,0 +1,2 @@
+CONTENT_TYPE_TEXT_CSV = "text/csv"
+IDENTIFIER = "_id"
